fix(source_provider): skip source link_type not match provider's link_type
